Move NullConfigActionHandler into a separate file
